ingress_server/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ingress_server"
version = "0.2.0"
description = "Backend-only ingestion service: HTTP endpoints that accept uploads and persist validated tables to an S3-backed Zarr store (no UI)."
readme = "README.md"
urls = { Homepage = "https://github.com/geomop/zarr_fuse" }
authors = [{ name = "Jan Brezina", email = "jan.brezina@tul.cz" }]
license = { text = "BSD-3-Clause" }
requires-python = ">=3.11"
keywords = [
  "backend",
  "ingestion",
  "upload",
  "HTTP API",
  "S3",
  "Zarr",
  "fsspec",
  "Flask",
  "CSV",
  "data pipeline",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "PyYAML",
  "flask",
  "Flask-HTTPAuth",
  "python-dotenv",
  "polars",
  "requests",
  "apscheduler",
  "pydantic",
  "xarray",
  "cfgrib",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
ingress_server = "ingress_server.main:main"
